Make BTC also subject to 'max 1% of coin amount lost in transfer' heuristic
Right now the BTC transfer fee on Bittrex is really high (0.001 BTC)
and without this, the default min-btc-xfer setting of (0.025) causes
us to loose too much BTC during each trasfer.

We currently apply a heuristic to altcoin transfers to avoid losing
more than 1% of them in transfers.  There is no reason that BTC should
be exempt from this policy.