Add some protection against old installations with stale PHP files
When the bot PHP subdirectory was created, some old installations may be
left with stranding old copies of some files in the main installation
subfolder which may be imported at unexpected times.  This code helps
detect them explicitly and refuse to run the bot.  We also tell the user
which files need to be deleted to fix the problem.