Add an extra trade check to ensure trade safety with buy/sell-rate
Previously we would only check that the sell rate is
above the buy rate before deciding to trade.  But if the
user has set the buy-rate and sell-rate config variables,
they can affect the actual rate at which the limit orders
are placed, so we need to perform an  extra check right
before placing orders as well.