Revert "Take the amount of altcoins at the source exchange into account"
This reverts commit 531bbfc3f8373e745daa24b0b7ef2ff0ae316f2d.

This commit introduced a bug where if we only have the altcoin balance
on the target exchange, the bot will become unable to leverage arbitrage
opportunities that arise on exchanges where we have a zero balance of
that altcoin.

See https://gitter.im/cryptoeax-arbbot/Lobby?at=5a553a1283152df26d4fe203
for the discussion.